Extract EvalNotes and GradesReminder component from new eval
