+++ src/Pages/Eval/NewStudentEval.jsx
@@ -15,43 +15,29 @@ import { useNavigate, useParams } from "react-router-dom";
 
 import dayjs from "dayjs";
 import { ref, uploadBytes } from "firebase/storage";
-import {
-  Button,
-  Card,
-  Collapse,
-  Container,
-  Modal,
-  OverlayTrigger,
-  Popover,
-} from "react-bootstrap";
+import { Button, OverlayTrigger, Popover } from "react-bootstrap";
 import { auth, db, storage } from "../../Services/firebase";
 import { ToastContext } from "../../Services/toast";
 import EvalFooter from "./Components/EvalFooter";
 import EvalHeader from "./Components/EvalHeader";
 import Tasks from "./Components/Tasks";
 import StandardInfo from "../Standards/Components/StandardInfo";
+import EvalNotes from "./Components/EvalNotes";
+import GradesReminder from "./Components/GradesReminder";
 
 const NewStudentEval = () => {
-  const [standards, setStandards] = useState([]);
+  const [standardOptions, setStandardOptions] = useState([]);
+
+  const [recentStandards, setRecentStandards] = useState([]);
   const [standardAverages, setStandardAverages] = useState({});
   const [standardSuggestions, setStandardSuggestions] = useState([]);
-  const [standardOptions, setStandardOptions] = useState([]);
 
   const [showStandardInfo, setShowStandardInfo] = useState(false);
   const [selectedStandard, setSelectedStandard] = useState(null);
 
   const [loading, setLoading] = useState(true);
 
-  const [notes, setNotes] = useState([]);
-  const [notesIndex, setNotesIndex] = useState(0);
-  const [showNotes, setShowNotes] = useState(false);
-
-  const [gradesReminderMessage, setGradesReminderMessage] = useState("");
-  const [studentOverElementary, setStudentOverElementary] = useState(false);
-  const [showGradesReminder, setShowGradesReminder] = useState(false);
-
-  const [gradesTooltip, setGradesTooltip] = useState("");
-  const [showGradesTooltip, setShowGradesTooltip] = useState(false);
+  const [recentEvals, setRecentEvals] = useState([]);
 
   const addToast = useContext(ToastContext);
 
@@ -94,6 +80,7 @@ const NewStudentEval = () => {
     window.scrollTo(0, 0);
   }, []);
 
+  // prep eval with student data
   useEffect(() => {
     const unsubscribeStudents = onSnapshot(
       doc(db, "students", params.studentid),
@@ -114,59 +101,7 @@ const NewStudentEval = () => {
     };
   }, [params.studentid]);
 
-  useEffect(() => {
-    getDoc(studentRef.current).then((student) => {
-      if (parseInt(student.data().student_grade) >= 6) {
-        setStudentOverElementary(true);
-      } else {
-        setStudentOverElementary(false);
-      }
-    });
-  }, [studentRef]);
-
-  useEffect(() => {
-    if (!evaluation.student_id || !evaluation.student_name) return;
-
-    const unsubscribeGrades = onSnapshot(
-      query(
-        collection(db, "grades"),
-        where("student_id", "==", evaluation.student_id),
-        orderBy("date", "desc"),
-        limit(1),
-      ),
-      (gradesSnapshot) => {
-        if (gradesSnapshot.docs.length > 0) {
-          const grade = gradesSnapshot.docs[0].data();
-          if (dayjs(grade.date).isBefore(dayjs().subtract(2, "week"))) {
-            setGradesReminderMessage(
-              `${evaluation.student_name} has not had their class grades updated since ${dayjs(
-                grade.date,
-              ).format("MMMM D, YYYY")}.`,
-            );
-            setGradesTooltip(
-              `We keep track of student progress and performance by recording their class grades every two weeks.`,
-            );
-            setShowGradesReminder(true);
-          }
-        } else {
-          if (studentOverElementary) {
-            setGradesReminderMessage(
-              `${evaluation.student_name} has not had their class grades entered yet.`,
-            );
-            setGradesTooltip(
-              "Students in 6th grade and above are expected to have their class grades entered regularly to track progress and performance.",
-            );
-            setShowGradesReminder(true);
-          }
-        }
-      },
-    );
-
-    return () => {
-      unsubscribeGrades();
-    };
-  }, [evaluation.student_id, evaluation.student_name, studentOverElementary]);
-
+  // prep eval with tutor data
   useEffect(() => {
     if (evaluation.tutor_id) {
       getDoc(doc(db, "tutors", evaluation.tutor_id)).then((tutor) => {
@@ -180,6 +115,7 @@ const NewStudentEval = () => {
     }
   }, [evaluation.tutor_id]);
 
+  // load cached eval data
   useEffect(() => {
     if (localStorage.getItem(`${params.studentid}_eval`)) {
       setEvaluation(
@@ -221,108 +157,69 @@ const NewStudentEval = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [params.studentid]);
 
+  // load recent evals for notes
   useEffect(() => {
-    const evalsQuery = query(
-      collection(db, "evaluations"),
-      where("student_id", "==", params.studentid),
-      orderBy("date", "desc"),
-      limit(5),
+    const unsubscribeRecentEvals = onSnapshot(
+      query(
+        collection(db, "evaluations"),
+        where("student_id", "==", params.studentid),
+        orderBy("date", "desc"),
+        limit(5),
+      ),
+      (evalsSnapshot) => {
+        const taskPromises = evalsSnapshot.docs.map(async (e) => {
+          return {
+            ...e.data(),
+            id: e.id,
+            tasks: await getDocs(collection(e.ref, "tasks")).then(
+              (taskSnapshot) => {
+                return taskSnapshot.docs.map((task) => {
+                  return {
+                    ...task.data(),
+                    id: task.id,
+                  };
+                });
+              },
+            ),
+          };
+        });
+
+        Promise.all(taskPromises).then((evals) => {
+          setRecentEvals(evals);
+        });
+      },
     );
-
-    const unsubscribeRecentEvals = onSnapshot(evalsQuery, (evalsSnapshot) => {
-      const notesArray = evalsSnapshot.docs.map((e) => {
-        return {
-          tutor: e.data().tutor_name,
-          date: e.data().date,
-          notes: e.data().next_session,
-          id: e.id,
-        };
-      });
-
-      setNotes(notesArray);
-
-      const fetchTasksPromises = evalsSnapshot.docs.map(async (evaluation) => {
-        return getDocs(collection(evaluation.ref, "tasks")).then(
-          (tasksSnapshot) => {
-            const tasks = tasksSnapshot.docs.map((doc) => {
-              return {
-                ...doc.data(),
-                id: doc.id,
-              };
-            });
-
-            const fetchStandardsPromises = tasks.map((task) => {
-              const standardsPromises =
-                task.standards?.map(async (standard) => {
-                  if (standard === "") return Promise.resolve(null);
-                  return getDoc(
-                    doc(db, "standards", standard?.id || standard),
-                  ).then((sdata) => {
-                    return {
-                      ...sdata.data(),
-                      id: sdata.id,
-                      asof: evaluation.data().date,
-                    };
-                  });
-                }) || [];
-              return Promise.all(standardsPromises);
-            });
-
-            return Promise.all(fetchStandardsPromises);
-          },
-        );
-      });
-
-      Promise.all(fetchTasksPromises).then((standardsArray) => {
-        const flattenedStandards = standardsArray
-          .flat()
-          .flat()
-          .filter((s) => s !== null);
-        const uniqueStandards = flattenedStandards.reduce((acc, standard) => {
-          const existingStandard = acc.find((s) => s.key === standard.key);
-          if (!existingStandard) {
-            acc.push(standard);
-          }
-          return acc;
-        }, []);
-        // console.log(uniqueStandards);
-        setStandards(uniqueStandards);
-      });
-    });
 
     return () => {
       unsubscribeRecentEvals();
     };
   }, [params.studentid]);
 
+  // get all standards from recent evaluations
   useEffect(() => {
-    for (let i = 0; i < tasks.length; i++) {
-      if (
-        tasks[i].standards.length < 1 &&
-        tasks[i].progression !== "" &&
-        tasks[i].progression <= 2
-      ) {
-        document.getElementById("flagForReview").classList.remove("d-none");
-        return;
-      }
-      if (tasks[i].engagement !== "" && tasks[i].engagement <= 2) {
-        document.getElementById("flagForReview").classList.remove("d-none");
-        return;
-      }
-      for (let j = 0; j < tasks[i].standards.length; j++) {
-        if (
-          tasks[i].standards[j].progression &&
-          tasks[i].standards[j].progression <= 2
-        ) {
-          document.getElementById("flagForReview").classList.remove("d-none");
-          return;
-        }
-      }
-    }
+    let standards = [];
+    recentEvals.forEach((evaluation) => {
+      evaluation.tasks.forEach((task) => {
+        task.standards.forEach((standard) => {
+          if (!standards.includes(standard.id)) {
+            standards.push(standard.id);
+          }
+        });
+      });
+    });
 
-    document.getElementById("flagForReview").classList.add("d-none");
-  }, [tasks]);
+    Promise.all(
+      standards.map(async (s) => {
+        return getDoc(doc(db, "standards", s)).then((standard) => {
+          return { ...standard.data(), id: standard.id };
+        });
+      }),
+    ).then((standards) => {
+      setRecentStandards(standards);
+    });
+  }, [recentEvals]);
 
+  // calculate average progression for each standard across ALL evals
   useEffect(() => {
     const unsubscribeEvaluations = onSnapshot(
       query(
@@ -382,11 +279,12 @@ const NewStudentEval = () => {
     };
   }, [params.studentid]);
 
+  // generate suggestions based on average progression and recent standards
   useEffect(() => {
     let suggestions = [];
     let postreqmap = {};
     let suggestionPromises = [];
-    standards.forEach((standard) => {
+    recentStandards.forEach((standard) => {
       if (parseFloat(standardAverages[standard.id]) < 3.5) {
         suggestions.push({
           ...standard,
@@ -421,7 +319,36 @@ const NewStudentEval = () => {
         setStandardOptions(suggestions);
         setStandardSuggestions(suggestions);
       });
-  }, [standardAverages, standards]);
+  }, [standardAverages, recentStandards]);
+
+  // flag for review button
+  useEffect(() => {
+    for (let i = 0; i < tasks.length; i++) {
+      if (
+        tasks[i].standards.length < 1 &&
+        tasks[i].progression !== "" &&
+        tasks[i].progression <= 2
+      ) {
+        document.getElementById("flagForReview").classList.remove("d-none");
+        return;
+      }
+      if (tasks[i].engagement !== "" && tasks[i].engagement <= 2) {
+        document.getElementById("flagForReview").classList.remove("d-none");
+        return;
+      }
+      for (let j = 0; j < tasks[i].standards.length; j++) {
+        if (
+          tasks[i].standards[j].progression &&
+          tasks[i].standards[j].progression <= 2
+        ) {
+          document.getElementById("flagForReview").classList.remove("d-none");
+          return;
+        }
+      }
+    }
+
+    document.getElementById("flagForReview").classList.add("d-none");
+  }, [tasks]);
 
   function handleEvalChange(newEval) {
     setEvaluation(newEval);
@@ -443,6 +370,8 @@ const NewStudentEval = () => {
       });
     }
     let clean = true;
+
+    // validity checks
 
     if (!evaluation.tutor_id) {
       document.getElementById("tutor").classList.add("is-invalid");
@@ -607,40 +536,15 @@ const NewStudentEval = () => {
     }
   }
 
-  function color(progression) {
-    const parsed = parseFloat(progression);
-    let color;
-    if (parsed >= 3.5) {
-      color = "success";
-    } else if (parsed >= 2.5) {
-      color = "primary text-dark ";
-    } else if (parsed >= 1.5) {
-      color = "warning";
-    } else {
-      color = "danger";
-    }
-    return color;
-  }
-
   return (
     <>
       <div className='p-3 d-flex flex-column'>
         <div className='d-flex'>
           <h1 className='display-1'>New Session Evaluation</h1>
-          {notes[0] && (
-            <Button
-              variant=''
-              className='w-25 ms-auto'
-              onClick={() => setShowNotes(true)}
-            >
-              <Card className='shadow' style={{ cursor: "pointer" }}>
-                <Card.Header>Last Session's Notes</Card.Header>
-                <Card.Body>
-                  <div className='text-truncate'>{notes[0].notes}</div>
-                </Card.Body>
-              </Card>
-            </Button>
-          )}
+          <EvalNotes
+            recentEvals={recentEvals}
+            standardSuggestions={standardSuggestions}
+          />
         </div>
         <div className='d-flex flex-fill card p-3 m-3 bg-light-subtle'>
           <EvalHeader
@@ -743,311 +647,7 @@ const NewStudentEval = () => {
           </div>
         </div>
       </div>
-      <div className='d-flex'>
-        <Modal
-          show={showNotes}
-          onHide={() => setShowNotes(false)}
-          centered
-          size='lg'
-        >
-          <Modal.Header>
-            <Modal.Title>For Today's Session</Modal.Title>
-            <Button
-              variant='secondary'
-              onClick={() => setShowNotes(false)}
-              style={{ "--bs-bg-opacity": "0" }}
-            >
-              <i className='bi bi-x-lg' />
-            </Button>
-          </Modal.Header>
-          <Modal.Body>
-            <Card className='d-flex flex-column bg-light-subtle'>
-              <Card.Header>
-                <div className='h4'>Previous Session Notes</div>
-                <Card.Subtitle className='text-muted'>
-                  {evaluation.student_name}
-                </Card.Subtitle>
-              </Card.Header>
-              <Card.Body>
-                <div className='d-flex'>
-                  <div className='d-flex flex-column'>
-                    <div className='h6'>{notes[notesIndex]?.tutor}</div>
-                    <div className='text-muted'>
-                      {dayjs(notes[notesIndex]?.date).format("MMMM D, YYYY")}
-                    </div>
-                  </div>
-                  <div className='ms-auto align-self-center'>
-                    <Button
-                      variant='primary'
-                      size='sm'
-                      onClick={() => navigate(`/eval/${notes[notesIndex]?.id}`)}
-                    >
-                      View Evaluation{" "}
-                      <i className='ms-auto ps-1 bi bi-box-arrow-up-right'></i>
-                    </Button>
-                  </div>
-                </div>
-                <hr />
-                <div className=''>{notes[notesIndex]?.notes}</div>
-              </Card.Body>
-              <Card.Footer className='d-flex'>
-                <Button
-                  variant='primary'
-                  size='sm'
-                  className='me-auto'
-                  disabled={notesIndex >= notes.length - 1}
-                  onClick={() => {
-                    setNotesIndex(notesIndex + 1);
-                  }}
-                >
-                  <i className='bi bi-arrow-left' />
-                </Button>
-                <div className='text-muted align-self-center'>
-                  {notes.length - notesIndex} / {notes.length}
-                </div>
-                <Button
-                  variant='primary'
-                  size='sm'
-                  className='ms-auto'
-                  disabled={notesIndex <= 0}
-                  onClick={() => {
-                    setNotesIndex(notesIndex - 1);
-                  }}
-                >
-                  <i className='bi bi-arrow-right' />
-                </Button>
-              </Card.Footer>
-            </Card>
-            {standardSuggestions.length > 0 && (
-              <div className='w-100'>
-                <hr />
-                <Card className='d-flex flex-column bg-light-subtle'>
-                  <Card.Header>
-                    <div className='h4'>Standards</div>
-                    <Card.Subtitle className='text-muted'>
-                      Suggestions for {evaluation.student_name}
-                    </Card.Subtitle>
-                  </Card.Header>
-                  <Card.Body>
-                    <Container>
-                      <div className='row cols-auto justify-content-between'>
-                        {standardSuggestions.filter((s) => s.progression)
-                          .length > 0 && (
-                          <Card className='col'>
-                            <Card.Body>
-                              <div className='d-flex justify-content-between'>
-                                <h4>To Review</h4>
-                                <OverlayTrigger
-                                  placement='top'
-                                  overlay={
-                                    <Popover>
-                                      <Popover.Header>To Review</Popover.Header>
-                                      <Popover.Body>
-                                        These standards have an average
-                                        progression below 3.5 and have been
-                                        worked on recently.
-                                      </Popover.Body>
-                                    </Popover>
-                                  }
-                                >
-                                  <i className='bi bi-question-square'></i>
-                                </OverlayTrigger>
-                              </div>
-                              <div className='d-flex flex-wrap'>
-                                {standardSuggestions
-                                  .filter((s) => s.progression)
-                                  .map((standard) => (
-                                    <Card
-                                      className='bg-light-subtle w-100 m-1'
-                                      key={standard.id}
-                                    >
-                                      <Card.Body>
-                                        <div className='d-flex justify-content-between align-items-center'>
-                                          <Button
-                                            variant='link'
-                                            onClick={() => {
-                                              setSelectedStandard(standard);
-                                              setShowStandardInfo(true);
-                                            }}
-                                            style={{
-                                              "--bs-btn-padding-x": "0",
-                                              "--bs-btn-padding-y": "0",
-                                            }}
-                                          >
-                                            {standard.key}
-                                          </Button>
-                                          <div
-                                            className={`badge bg-${color(standard.progression)}`}
-                                          >
-                                            {standard.progression}
-                                          </div>
-                                        </div>
-                                        <div className='text-muted'>
-                                          As of{" "}
-                                          {dayjs(standard.asof).format(
-                                            "MMMM D, YYYY",
-                                          )}
-                                        </div>
-                                      </Card.Body>
-                                    </Card>
-                                  ))}
-                              </div>
-                            </Card.Body>
-                          </Card>
-                        )}
-                        {standardSuggestions.filter(
-                          (s) => s.progression === undefined,
-                        ).length > 0 && (
-                          <Card className='col'>
-                            <Card.Body>
-                              <div className='d-flex justify-content-between'>
-                                <h4>Up Next</h4>
-                                <OverlayTrigger
-                                  placement='top'
-                                  overlay={
-                                    <Popover>
-                                      <Popover.Header>Up Next</Popover.Header>
-                                      <Popover.Body>
-                                        These suggestions are based on standards
-                                        that have been mastered recently.
-                                      </Popover.Body>
-                                    </Popover>
-                                  }
-                                >
-                                  <i className='bi bi-question-square'></i>
-                                </OverlayTrigger>
-                              </div>
-                              <div className='d-flex flex-wrap'>
-                                {standardSuggestions
-                                  .filter((s) => s.progression === undefined)
-                                  .map((standard) => (
-                                    <Card
-                                      className='bg-light-subtle w-100 m-1'
-                                      key={standard.id}
-                                    >
-                                      <Card.Body className='d-flex justify-content-between align-items-center'>
-                                        <div className='d-flex flex-column'>
-                                          <div className='d-flex justify-content-between align-items-center'>
-                                            <Button
-                                              variant='link'
-                                              onClick={() => {
-                                                setSelectedStandard(
-                                                  standard.parent,
-                                                );
-                                                setShowStandardInfo(true);
-                                              }}
-                                              style={{
-                                                "--bs-btn-padding-x": "0",
-                                                "--bs-btn-padding-y": "0",
-                                              }}
-                                            >
-                                              {standard.parent.key}
-                                            </Button>
-                                            <span
-                                              className={`badge bg-${color(standard.parent.progression)}`}
-                                            >
-                                              {standard.parent.progression}
-                                            </span>
-                                          </div>
-                                          <div className='text-muted'>
-                                            As of{" "}
-                                            {dayjs(standard.parent.asof).format(
-                                              "MMMM D, YYYY",
-                                            )}
-                                          </div>
-                                        </div>
-                                        <i className='bi bi-arrow-right fs-2'></i>
-                                        <Button
-                                          variant='link'
-                                          className=''
-                                          onClick={() => {
-                                            setSelectedStandard(standard);
-                                            setShowStandardInfo(true);
-                                          }}
-                                          style={{
-                                            "--bs-btn-padding-x": "0",
-                                            "--bs-btn-padding-y": "0",
-                                          }}
-                                        >
-                                          {standard.key}
-                                        </Button>
-                                      </Card.Body>
-                                    </Card>
-                                  ))}
-                              </div>
-                            </Card.Body>
-                          </Card>
-                        )}
-                      </div>
-                    </Container>
-                  </Card.Body>
-                </Card>
-              </div>
-            )}
-          </Modal.Body>
-        </Modal>
-        <Modal
-          show={showGradesReminder}
-          onHide={() => setShowGradesReminder(false)}
-        >
-          <Modal.Header>
-            <Modal.Title>Grades Reminder</Modal.Title>
-            <Button
-              variant='secondary'
-              onClick={() => setShowGradesReminder(false)}
-              style={{ "--bs-bg-opacity": "0" }}
-            >
-              <i className='bi bi-x-lg' />
-            </Button>
-          </Modal.Header>
-          <Modal.Body>
-            <div className='d-flex flex-column'>
-              <div>{gradesReminderMessage}</div>
-              <hr />
-              <div>
-                Please make sure to update their grades before they leave.
-              </div>
-              <Button
-                variant='link'
-                size='sm'
-                className='me-auto link-secondary fst-italic'
-                style={{ "--bs-btn-padding-x": "0rem" }}
-                onClick={() => setShowGradesTooltip(!showGradesTooltip)}
-              >
-                Why am I seeing this?
-              </Button>
-              <Collapse in={showGradesTooltip}>
-                <div>
-                  {gradesTooltip}
-                  <br />
-                  If you believe this is a mistake or need an exception, please
-                  contact an administrator.
-                </div>
-              </Collapse>
-            </div>
-          </Modal.Body>
-          <Modal.Footer className='d-flex'>
-            <Button
-              variant='secondary'
-              size='sm'
-              className='me-auto'
-              onClick={() => setShowGradesReminder(false)}
-            >
-              I'll do it later
-            </Button>
-            <Button
-              variant='primary'
-              size='sm'
-              onClick={() => {
-                localStorage.setItem("student_tab", "grades");
-                navigate(`/students/${params.studentid}`);
-              }}
-            >
-              Take me there now
-            </Button>
-          </Modal.Footer>
-        </Modal>
-      </div>
+      <GradesReminder studentid={params.studentid} />
       <StandardInfo
         show={showStandardInfo}
         setShow={setShowStandardInfo}
